# out/openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: "The Jobs API allows you to create, edit, and delete jobs."
  title: Jobs API 2.1
  version: "2.1"
servers:
- description: "In the URL, substitute `<databricks-instance>` with the domain name\
    \ of your deployment. Use the form `<account>.cloud.databricks.com`."
  url: https://<databricks-instance>/api/
paths:
  /2.0/jobs/runs/export:
    get:
      description: Export and retrieve the job run task.
      operationId: jobs_runs_export
      parameters:
      - description: The canonical identifier for the run. This field is required.
        explode: true
        in: query
        name: run_id
        required: true
        schema:
          example: 455644833
          format: int64
          type: integer
        style: form
      - description: "Which views to export (CODE, DASHBOARDS, or ALL). Defaults to\
          \ CODE."
        explode: true
        in: query
        name: views_to_export
        required: false
        schema:
          $ref: '#/components/schemas/ViewsToExport'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsRunsExport_200_response'
          description: Run was exported successfully.
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Export and retrieve a job run
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/create:
    post:
      description: Create a new job.
      operationId: jobs_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobsCreate_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsCreate_200_response'
          description: Job was created successfully
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Create a new job
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/delete:
    post:
      description: Deletes a job.
      operationId: jobs_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobsDelete_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties: {}
                type: object
          description: Job was deleted successfully.
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Delete a job
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/get:
    get:
      description: Retrieves the details for a single job.
      operationId: jobs_get
      parameters:
      - description: The canonical identifier of the job to retrieve information about.
          This field is required.
        explode: true
        in: query
        name: job_id
        required: true
        schema:
          example: 11223344
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsGet_200_response'
          description: Job was retrieved successfully.
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Get a single job
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/list:
    get:
      description: Retrieves a list of jobs.
      operationId: jobs_list
      parameters:
      - description: The number of jobs to return. This value must be greater than
          0 and less or equal to 25. The default value is 20.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          example: 25
          maximum: 25
          minimum: 1
          type: integer
        style: form
      - description: "The offset of the first job to return, relative to the most\
          \ recently created job."
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          example: 0
          minimum: 0
          type: integer
        style: form
      - description: Whether to include task and cluster details in the response.
        explode: true
        in: query
        name: expand_tasks
        required: false
        schema:
          default: false
          example: false
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsList_200_response'
          description: List of jobs was retrieved successfully.
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: List all jobs
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/reset:
    post:
      description: Overwrites all the settings for a specific job. Use the Update
        endpoint to update job settings partially.
      operationId: jobs_reset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobsReset_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties: {}
                type: object
          description: Job was overwritten successfully.
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Overwrites all settings for a job
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/run-now:
    post:
      description: Run a job and return the `run_id` of the triggered run.
      operationId: jobs_run_now
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobsRunNow_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsRunNow_200_response'
          description: Run was started successfully.
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Trigger a new job run
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/runs/cancel:
    post:
      description: "Cancels a run. The run is canceled asynchronously, so when this\
        \ request completes, the run may still be running. The run are terminated\
        \ shortly. If the run is already in a terminal `life_cycle_state`, this method\
        \ is a no-op."
      operationId: jobs_runs_cancel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobsRunsCancel_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties: {}
                type: object
          description: Run was cancelled successfully.
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Cancel a job run
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/runs/delete:
    post:
      description: Deletes a non-active run. Returns an error if the run is active.
      operationId: jobs_runs_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobsRunsDelete_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties: {}
                type: object
          description: Run was deleted successfully.
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Delete a job run
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/runs/get:
    get:
      description: Retrieve the metadata of a run.
      operationId: jobs_runs_get
      parameters:
      - description: The canonical identifier of the run for which to retrieve the
          metadata. This field is required.
        explode: true
        in: query
        name: run_id
        required: true
        schema:
          example: 455644833
          format: int64
          type: integer
        style: form
      - description: Whether to include the repair history in the response.
        explode: true
        in: query
        name: include_history
        required: false
        schema:
          example: true
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsRunsGet_200_response'
          description: Run was retrieved successfully
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Get a single job run
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /2.1/jobs/runs/get-output:
    get:
      description: "Retrieve the output and metadata of a run. When a notebook task\
        \ returns a value through the dbutils.notebook.exit() call, you can use this\
        \ endpoint to retrieve that value. Databricks restricts this API to return\
        \ the first 5 MB of the output. To return a larger result, you can store job\
        \ results in a cloud storage service.\nThis endpoint validates that the run_id\
        \ parameter is valid and returns an HTTP status code 400 if the run_id parameter\
        \ is invalid.\nRuns are automatically removed after 60 days. If you to want\
        \ to reference them beyond 60 days, you must save old run results before they\
        \ expire. To export using the UI, see Export job run results. To export using\
        \ the Jobs API, see Runs export."
      operationId: jobs_runs_get_output
      parameters:
      - description: The canonical identifier for the run. This field is required.
        explode: true
        in: query
        name: run_id
        required: true
        schema:
          example: 455644833
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsRunsGetOutput_200_response'
          description: Run output was retrieved successfully.
        "400":
          content:
            application/json:
              example:
                error_code: INVALID_PARAMETER_VALUE
                message: Invalid value for parameter job_id
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was malformed. See JSON response for error details.
        "401":
          content:
            application/json:
              example:
                error_code: PERMISSION_DENIED
                message: Unauthorized access.
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was unauthorized.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request was not handled correctly due to a server error.
      security:
      - bearerAuth: []
      summary: Get the output for a single run
      x-openapi-router-controller: openapi_server.controllers.default_controller
...
components:
...
  schemas:
...
    ViewsToExport:
      default: CODE
      description: "* `CODE`: Code view of the notebook.\n* `DASHBOARDS`: All dashboard\
        \ views of the notebook.\n* `ALL`: All views of the notebook."
      enum:
      - CODE
      - DASHBOARDS
      - ALL
      title: ViewsToExport
      type: string
...
    Error:
      properties:
        error_code:
          description: Error code
          example: INTERNAL_ERROR
          title: error_code
          type: string
        message:
          description: Human-readable error message that describes the cause of the
            error.
          example: Unexpected error.
          title: message
          type: string
      title: Error
      type: object
...
    JobsCreate_request:
      allOf:
      - $ref: '#/components/schemas/JobSettings'
      - $ref: '#/components/schemas/AccessControlList'
      title: JobsCreate_request
    JobsCreate_200_response:
      example:
        job_id: 11223344
      properties:
        job_id:
          description: The canonical identifier for the newly created job.
          example: 11223344
          format: int64
          title: job_id
          type: integer
      title: JobsCreate_200_response
      type: object
...
    JobsDelete_request:
      properties:
        job_id:
          description: The canonical identifier of the job to delete. This field is
            required.
          example: 11223344
          format: int64
          title: job_id
          type: integer
      required:
      - job_id
      title: JobsDelete_request
      type: object
...
    JobsRunsExport_200_response:
      example:
        views:
        - name: name
          type: null
          content: content
        - name: name
          type: null
          content: content
      properties:
        views:
          description: The exported content in HTML format (one for every view item).
          items:
            $ref: '#/components/schemas/ViewItem'
          title: views
          type: array
      title: JobsRunsExport_200_response
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: api_token
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller_.info_from_bearerAuth